Converted `SingleUseTokenProvider` to TypeScript
no issue

- following our scout rule that any files touched should be converted to TypeScript
